Add a .clang-fomat file
This adds a clang-format file and a modification of the existing
openssl style guide corresponding to it in STYLE.md. The
changes to STYLE.md include removal of much of the whitespace
specifics that clang-format would enforce, and a few clarifications
on include files and modern code vs legacy code. Probably
needs a read through