Add a STYLE.md file to document where we are going
This puts the style guide into the repository which is sort of
the modern "norm". This has taken the text from the existing
style guide in the web page and adapted it to how things
will be with an automatic clang-format enforcement.

What I have done is:

- Referenced the Linux kernel style guide as the base we are using
including the link to it.

- Removed any reference to spacing, indenting, continuations, etc
etc. which are contained in the Linux kernel style guide and
enforced by clang-format.

- Added a section on Integers.

- Added a section for "new code" to Function return values
  preserving the existing advice for legacy code.

- Added a secion on Header Files
  As I believe we can move to cleaner header files soon, this
  merely documents the expectation that I believe we can get to
  (clang-format does not yet do any header re-ordering)