Add find_sorted to stack.
find_sorted is just like find, but requires that the stack
be sorted.  It will not mutate the stack by sorting it if it
is not, but rather will fail with ossl_assert, returning -1
in a production build if the stack is not sorted.

We have a number of places where we want to find things
in stacks under lock.