Make ASN1_BIT_STRING_set set the number of bits used.
And do not implicitly truncate on output.

Formerly we would always attempt to truncate the bit string
when writing it. This ensures it stays consistant.

The gotcha here is that if someone constructs a bit string
without using the BIT_STRING_set functions, but rather
does ASN1_STRING_set(bitstr, yolo_bytes); This could produce
an incorrect encoding with trailing 0's if they were relying
on the implicit truncation on output behavior before.